Factor out signing test helpers
